Xml samuel

+++ app/src/main/java/com/cgi/sdm_project/logica/juego/juego/selectores/ProbabilitySelector.java
@@ -6,7 +6,7 @@ public class ProbabilitySelector extends BasicSelector {
     @Override
     public Reglas getSiguienteJuego() {
         int index = 0;
-        int probabilidad = (int) (Math.random() * 100);
+        float probabilidad = (float) (Math.random() * 100);
         while ((probabilidad -= getElementAt(index).probabilidad()) > 0)
             index++;
         return getElementAt(index);
